chore: rename todoreducer to note reducer
