+++ src/reducers/noteReducer.js
